+++ NPLRuntime/ParaEngine/3dengine/Win32MCI.h
@@ -1,0 +1,15 @@
+#pragma once
+
+#include "MCIController.h"
+
+class Win32MCI : public ParaEngine::MCIController::Interface
+{
+public:
+	void Release();
+
+	bool Start();
+
+	bool Stop();
+
+	bool Save(const char* fileName);
+};
